+++ model/gurps/general_settings.go
@@ -61,7 +61,7 @@ const (
 	MaximumAutoColWidthDef     = 800
 )
 
-const currentGeneralSettingsVersion = 1
+const currentGeneralSettingsVersion = 2
 
 // GeneralSettings holds general settings for a sheet.
 type GeneralSettings struct {
@@ -91,6 +91,7 @@ type GeneralSettings struct {
 	GroupContainersOnSort       bool             `json:"group_containers_on_sort"`
 	InitialFieldClickSelectsAll bool             `json:"initial_field_click_selects_all"`
 	RestoreWorkspaceOnStart     bool             `json:"restore_workspace_on_start"`
+	ExpandPageReferences        bool             `json:"expand_page_references"`
 }
 
 // NewGeneralSettings creates settings with factory defaults.
@@ -117,6 +118,7 @@ func NewGeneralSettings() *GeneralSettings {
 		AutoFillProfile:            true,
 		AutoAddNaturalAttacks:      true,
 		RestoreWorkspaceOnStart:    true,
+		ExpandPageReferences:       true,
 	}
 }
 
@@ -178,8 +180,14 @@ func (s *GeneralSettings) MonitorPPI() int {
 // EnsureValidity checks the current settings for validity and if they aren't valid, makes them so.
 func (s *GeneralSettings) EnsureValidity() {
 	if s.Version != currentGeneralSettingsVersion {
-		// Adjust from old default of 150% to new default of 100%
-		s.InitialSheetUIScale = s.InitialSheetUIScale * 100 / 150
+		if s.Version < 1 {
+			// Adjust from old default of 150% to new default of 100%
+			s.InitialSheetUIScale = s.InitialSheetUIScale * 100 / 150
+		}
+		if s.Version < 2 {
+			// This setting didn't exist before, so enable it to match the new default behavior.
+			s.ExpandPageReferences = true
+		}
 		s.Version = currentGeneralSettingsVersion
 	}
 	s.InitialPoints = fxp.ResetIfOutOfRange(s.InitialPoints, InitialPointsMin, InitialPointsMax, InitialPointsDef)